I should probably spread out commits
Removed customweapon check because it was breaking weapons

Finally fixed attack speed

Changed rarity declaration of items

Prevented blocks from being exploded

Updated player scoreboard when they kill or break a block

Removed block drops from various blocks

+++ core/src/main/java/net/siegerpg/siege/core/listeners/NPC/RichardBanker.java
@@ -140,7 +140,7 @@ public class RichardBanker implements Listener {
         int bankAmt = PlayerBanking.bankAmounts.get(player);
         int upgradeCost = bankLvl*5000;
         short upgradedLvl = (short) (bankLvl+1);
-        double pocketBal = VaultHook.econ.getBalance(player);
+        int pocketBal = (int) VaultHook.econ.getBalance(player);
         int maxAmt = bankLvl*5000;
         int diff = (maxAmt-bankAmt);
         if (pocketBal < diff) {
